chore: adds new settings for routes and more

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -11,30 +11,44 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_type_id');
+                $table->foreign('user_type_id')->references('id')->on('user_types')->cascadeOnDelete();
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('password');
+                $table->string('cpf')->unique();
+                $table->string('phone', 16)->nullable();
+                $table->enum('professional_type', ['admin', 'manager','basic', 'primary', 'secondary', 'other'])->nullable();
+                $table->string('document', 35)->unique()->nullable(); // Professional registration document
+                $table->string('address')->nullable();
+                $table->string('city')->nullable();
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
 
-        Schema::create('password_reset_tokens', function (Blueprint $table) {
-            $table->string('email')->primary();
-            $table->string('token');
-            $table->timestamp('created_at')->nullable();
-        });
+        if(!Schema::hasTable('password_reset_tokens')) {
+            Schema::create('password_reset_tokens', function (Blueprint $table) {
+                $table->string('email')->primary();
+                $table->string('token');
+                $table->timestamp('created_at')->nullable();
+            });
+        }
 
-        Schema::create('sessions', function (Blueprint $table) {
-            $table->string('id')->primary();
-            $table->foreignId('user_id')->nullable()->index();
-            $table->string('ip_address', 45)->nullable();
-            $table->text('user_agent')->nullable();
-            $table->longText('payload');
-            $table->integer('last_activity')->index();
-        });
+        if(!Schema::hasTable('sessions')) {
+            Schema::create('sessions', function (Blueprint $table) {
+                $table->string('id')->primary();
+                $table->foreignId('user_id')->nullable()->index();
+                $table->string('ip_address', 45)->nullable();
+                $table->text('user_agent')->nullable();
+                $table->longText('payload');
+                $table->integer('last_activity')->index();
+            });
+        }
     }
 
     /**
@@ -42,8 +56,14 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
-        Schema::dropIfExists('password_reset_tokens');
-        Schema::dropIfExists('sessions');
+        if (Schema::hasTable('users')) {
+            Schema::dropIfExists('users');
+        }
+        if (Schema::hasTable('password_reset_tokens')) {
+            Schema::dropIfExists('password_reset_tokens');
+        }
+        if (Schema::hasTable('sessions')) {
+            Schema::dropIfExists('sessions');
+        }
     }
 };
